+++ ext/MatrixAlgebraKitGenericSchurExt.jl
@@ -21,15 +21,6 @@ function MatrixAlgebraKit.default_exponential_algorithm(
     return MatrixFunctionViaEig(eig_alg)
 end
 
-function MatrixAlgebraKit.default_squareroot_algorithm(
-        type::Type{T}; domain_atol = nothing, kwargs...
-    ) where {T <: StridedMatrix{<:GSFloat}}
-    # the remaining keywords configure the eigensolver, `domain_atol` the domain check
-    eig_alg = MatrixAlgebraKit.default_eig_algorithm(type; kwargs...)
-    return isnothing(domain_atol) ? MatrixFunctionViaEig(eig_alg) :
-        MatrixFunctionViaEig(eig_alg; domain_atol)
-end
-
 function geev!(::GS, A::AbstractMatrix, Dd::AbstractVector, V::AbstractMatrix; kwargs...)
     D, Vmat = GenericSchur.eigen!(A)
     copyto!(Dd, D)