+++ src/MatrixAlgebraKit.jl
@@ -42,7 +42,8 @@ export LAPACK_HouseholderQR, LAPACK_HouseholderLQ, LAPACK_Simple, LAPACK_Expert,
 export GLA_HouseholderQR, GLA_QRIteration, GS_QRIteration
 export LQViaTransposedQR
 export PolarViaSVD, PolarNewton
-export MatrixFunctionViaLA, MatrixFunctionViaEig, MatrixFunctionViaEigh, MatrixFunctionViaTaylor
+export MatrixFunctionViaLA, MatrixFunctionViaEig, MatrixFunctionViaEigh, MatrixFunctionViaTaylor,
+    MatrixFunctionViaSchur
 export DefaultAlgorithm
 export DiagonalAlgorithm
 export NativeBlocked
@@ -94,6 +95,7 @@ include("common/pullbacks.jl")
 include("common/safemethods.jl")
 include("common/view.jl")
 include("common/regularinv.jl")
+include("common/quasitriangular.jl")
 include("common/matrixproperties.jl")
 include("common/balancing.jl")
 include("common/utility.jl")